+++ src/com/landawn/abacus/util/IOUtil.java
@@ -3927,7 +3927,7 @@ public final class IOUtil {
                 iterators.add(new LineIterator(reader));
             }
 
-            N.parse(iterators, lineOffset, count, 0, processThreadNum, queueSize, lineParser, onComplete);
+            Iterables.parse(iterators, lineOffset, count, 0, processThreadNum, queueSize, lineParser, onComplete);
         } finally {
             for (Reader reader : readers) {
                 closeQuietly(reader);
@@ -4022,7 +4022,7 @@ public final class IOUtil {
                 iterators.add(new LineIterator(reader));
             }
 
-            N.parse(iterators, lineOffset, count, readThreadNum, processThreadNum, queueSize, lineParser, onComplete);
+            Iterables.parse(iterators, lineOffset, count, readThreadNum, processThreadNum, queueSize, lineParser, onComplete);
         } finally {
             for (Reader reader : readers) {
                 closeQuietly(reader);
@@ -4115,7 +4115,7 @@ public final class IOUtil {
     public static <E extends Exception, E2 extends Exception> void parse(final Reader reader, final long lineOffset, final long count,
             final int processThreadNum, final int queueSize, final Try.Consumer<String, E> lineParser, final Try.Runnable<E2> onComplete)
             throws UncheckedIOException, E, E2 {
-        N.parse(new LineIterator(reader), lineOffset, count, processThreadNum, queueSize, lineParser, onComplete);
+        Iterables.parse(new LineIterator(reader), lineOffset, count, processThreadNum, queueSize, lineParser, onComplete);
     }
 
     private static InputStream openFile(final Holder<ZipFile> outputZipFile, final File file) throws IOException {