+++ src/com/landawn/abacus/util/Iterables.java
@@ -34,6 +34,13 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.RandomAccess;
 import java.util.Set;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import com.landawn.abacus.logging.Logger;
+import com.landawn.abacus.logging.LoggerFactory;
+import com.landawn.abacus.util.stream.Stream;
 
 /**
  * It's an extension and wrapper for Google Guava.
@@ -43,6 +50,7 @@ import java.util.Set;
  * @author Haiyang Li 
  */
 public final class Iterables {
+    private static final Logger logger = LoggerFactory.getLogger(Iterables.class);
 
     private Iterables() {
         // singleton.
@@ -303,6 +311,20 @@ public final class Iterables {
         }
     }
 
+    public static <T, E extends Exception> OptionalInt findFirstIndex(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return OptionalInt.empty();
+        }
+
+        for (int len = a.length, i = 0; i < len; i++) {
+            if (predicate.test(a[i])) {
+                return OptionalInt.of(i);
+            }
+        }
+
+        return OptionalInt.empty();
+    }
+
     public static <T, E extends Exception> OptionalInt findFirstIndex(final Collection<? extends T> c, final Try.Predicate<? super T, E> predicate) throws E {
         if (N.isNullOrEmpty(c)) {
             return OptionalInt.empty();
@@ -316,6 +338,20 @@ public final class Iterables {
             }
 
             idx++;
+        }
+
+        return OptionalInt.empty();
+    }
+
+    public static <T, E extends Exception> OptionalInt findLastIndex(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return OptionalInt.empty();
+        }
+
+        for (int len = a.length, i = len - 1; i >= 0; i--) {
+            if (predicate.test(a[i])) {
+                return OptionalInt.of(i);
+            }
         }
 
         return OptionalInt.empty();
@@ -371,6 +407,20 @@ public final class Iterables {
         }
     }
 
+    public static <T, E extends Exception> Nullable<T> findFirst(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return Nullable.empty();
+        }
+
+        for (int len = a.length, i = 0; i < len; i++) {
+            if (predicate.test(a[i])) {
+                return Nullable.of(a[i]);
+            }
+        }
+
+        return Nullable.empty();
+    }
+
     public static <T, E extends Exception> Nullable<T> findFirst(final Collection<T> c, Try.Predicate<? super T, E> predicate) throws E {
         if (N.isNullOrEmpty(c)) {
             return Nullable.empty();
@@ -385,8 +435,36 @@ public final class Iterables {
         return Nullable.empty();
     }
 
+    public static <T, E extends Exception> Nullable<T> findLast(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return Nullable.empty();
+        }
+
+        for (int len = a.length, i = len - 1; i >= 0; i--) {
+            if (predicate.test(a[i])) {
+                return Nullable.of(a[i]);
+            }
+        }
+
+        return Nullable.empty();
+    }
+
     public static <T, E extends Exception> Nullable<T> findLast(final Collection<T> c, Try.Predicate<? super T, E> predicate) throws E {
         return findLast(c, predicate, false);
+    }
+
+    public static <T, E extends Exception> Optional<T> findFirstNonNull(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return Optional.empty();
+        }
+
+        for (int len = a.length, i = 0; i < len; i++) {
+            if (a[i] != null && predicate.test(a[i])) {
+                return Optional.of(a[i]);
+            }
+        }
+
+        return Optional.empty();
     }
 
     public static <T, E extends Exception> Optional<T> findFirstNonNull(final Collection<T> c, Try.Predicate<? super T, E> predicate) throws E {
@@ -397,6 +475,20 @@ public final class Iterables {
         for (T e : c) {
             if (e != null && predicate.test(e)) {
                 return Optional.of(e);
+            }
+        }
+
+        return Optional.empty();
+    }
+
+    public static <T, E extends Exception> Optional<T> findLastNonNull(final T[] a, final Try.Predicate<? super T, E> predicate) throws E {
+        if (N.isNullOrEmpty(a)) {
+            return Optional.empty();
+        }
+
+        for (int len = a.length, i = len - 1; i >= 0; i--) {
+            if (a[i] != null && predicate.test(a[i])) {
+                return Optional.of(a[i]);
             }
         }
 
@@ -1104,6 +1196,184 @@ public final class Iterables {
 
         private int getAxisIndexForProductIndex(int index, int axis) {
             return (index / axesSizeProduct[axis + 1]) % axes[axis].length;
+        }
+    }
+
+    public static <T, E extends Exception> void parse(final Iterator<? extends T> iter, final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iter, elementParser, Fn.emptyAction());
+    }
+
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Iterator<? extends T> iter, final Try.Consumer<? super T, E> elementParser,
+            final Try.Runnable<E2> onComplete) throws E, E2 {
+        parse(iter, 0, Long.MAX_VALUE, elementParser, onComplete);
+    }
+
+    public static <T, E extends Exception> void parse(final Iterator<? extends T> iter, final long offset, final long count,
+            final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iter, offset, count, elementParser, Fn.emptyAction());
+    }
+
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Iterator<? extends T> iter, final long offset, final long count,
+            final Try.Consumer<? super T, E> elementParser, final Try.Runnable<E2> onComplete) throws E, E2 {
+        parse(iter, offset, count, 0, 0, elementParser, onComplete);
+    }
+
+    public static <T, E extends Exception> void parse(final Iterator<? extends T> iter, long offset, long count, final int processThreadNum,
+            final int queueSize, final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iter, offset, count, processThreadNum, queueSize, elementParser, Fn.emptyAction());
+    }
+
+    /**
+     * Parse the elements in the specified iterators one by one.
+     * 
+     * @param iter
+     * @param offset
+     * @param count
+     * @param processThreadNum new threads started to parse/process the lines/records
+     * @param queueSize size of queue to save the processing records/lines loaded from source data. Default size is 1024.
+     * @param elementParser.
+     * @param onComplete
+     * @param onError
+     */
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Iterator<? extends T> iter, long offset, long count,
+            final int processThreadNum, final int queueSize, final Try.Consumer<? super T, E> elementParser, final Try.Runnable<E2> onComplete) throws E, E2 {
+        parse(N.asList(iter), offset, count, 0, processThreadNum, queueSize, elementParser, onComplete);
+    }
+
+    public static <T, E extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators,
+            final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iterators, elementParser, Fn.emptyAction());
+    }
+
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators,
+            final Try.Consumer<? super T, E> elementParser, final Try.Runnable<E2> onComplete) throws E, E2 {
+        parse(iterators, 0, Long.MAX_VALUE, elementParser, onComplete);
+    }
+
+    public static <T, E extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators, final long offset, final long count,
+            final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iterators, offset, count, elementParser, Fn.emptyAction());
+    }
+
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators, final long offset,
+            final long count, final Try.Consumer<? super T, E> elementParser, final Try.Runnable<E2> onComplete) throws E, E2 {
+        parse(iterators, offset, count, 0, 0, 0, elementParser, onComplete);
+    }
+
+    public static <T, E extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators, final int readThreadNum,
+            final int processThreadNum, final int queueSize, final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iterators, readThreadNum, processThreadNum, queueSize, elementParser, Fn.emptyAction());
+    }
+
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators,
+            final int readThreadNum, final int processThreadNum, final int queueSize, final Try.Consumer<? super T, E> elementParser,
+            final Try.Runnable<E2> onComplete) throws E {
+        parse(iterators, 0, Long.MAX_VALUE, readThreadNum, processThreadNum, queueSize, elementParser);
+    }
+
+    public static <T, E extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators, final long offset, final long count,
+            final int readThreadNum, final int processThreadNum, final int queueSize, final Try.Consumer<? super T, E> elementParser) throws E {
+        parse(iterators, offset, count, readThreadNum, processThreadNum, queueSize, elementParser, Fn.emptyAction());
+    }
+
+    /**
+     * Parse the elements in the specified iterators one by one.
+     * 
+     * @param iterators
+     * @param offset
+     * @param count
+     * @param readThreadNum new threads started to parse/process the lines/records
+     * @param processThreadNum new threads started to parse/process the lines/records
+     * @param queueSize size of queue to save the processing records/lines loaded from source data. Default size is 1024.
+     * @param elementParser.
+     * @param onComplete
+     */
+    public static <T, E extends Exception, E2 extends Exception> void parse(final Collection<? extends Iterator<? extends T>> iterators, final long offset,
+            final long count, final int readThreadNum, final int processThreadNum, final int queueSize, final Try.Consumer<? super T, E> elementParser,
+            final Try.Runnable<E2> onComplete) throws E, E2 {
+        N.checkArgument(offset >= 0 && count >= 0, "'offset'=%s and 'count'=%s can not be negative", offset, count);
+
+        if (N.isNullOrEmpty(iterators)) {
+            return;
+        }
+
+        if (logger.isInfoEnabled()) {
+            logger.info("### Start to parse");
+        }
+
+        try (final Stream<T> stream = ((readThreadNum > 0 || queueSize > 0)
+                ? Stream.parallelConcatt(iterators, (readThreadNum == 0 ? 1 : readThreadNum), (queueSize == 0 ? 1024 : queueSize))
+                : Stream.concatt(iterators))) {
+
+            final Iterator<? extends T> iteratorII = stream.skip(offset).limit(count).iterator();
+
+            if (processThreadNum == 0) {
+                while (iteratorII.hasNext()) {
+                    elementParser.accept(iteratorII.next());
+                }
+
+                if (onComplete != null) {
+                    onComplete.run();
+                }
+            } else {
+                final AtomicInteger activeThreadNum = new AtomicInteger();
+                final ExecutorService executorService = Executors.newFixedThreadPool(processThreadNum);
+                final Holder<Throwable> errorHolder = new Holder<>();
+
+                for (int i = 0; i < processThreadNum; i++) {
+                    activeThreadNum.incrementAndGet();
+
+                    executorService.execute(new Runnable() {
+                        @Override
+                        public void run() {
+                            T element = null;
+                            try {
+                                while (errorHolder.value() == null) {
+                                    synchronized (iteratorII) {
+                                        if (iteratorII.hasNext()) {
+                                            element = iteratorII.next();
+                                        } else {
+                                            break;
+                                        }
+                                    }
+
+                                    elementParser.accept(element);
+                                }
+                            } catch (Exception e) {
+                                synchronized (errorHolder) {
+                                    if (errorHolder.value() == null) {
+                                        errorHolder.setValue(e);
+                                    } else {
+                                        errorHolder.value().addSuppressed(e);
+                                    }
+                                }
+                            } finally {
+                                activeThreadNum.decrementAndGet();
+                            }
+                        }
+                    });
+                }
+
+                while (activeThreadNum.get() > 0) {
+                    N.sleep(1);
+                }
+
+                if (errorHolder.value() == null && onComplete != null) {
+                    try {
+                        onComplete.run();
+                    } catch (Exception e) {
+                        errorHolder.setValue(e);
+                    }
+                }
+
+                if (errorHolder.value() != null) {
+                    throw N.toRuntimeException(errorHolder.value());
+                }
+            }
+        } finally {
+            if (logger.isInfoEnabled()) {
+                logger.info("### End to parse");
+            }
         }
     }
 }